src/app/button/button.component.spec.ts
import {
  async,
  ComponentFixture,
  TestBed
} from '@angular/core/testing';
import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
import { By } from '@angular/platform-browser';
import { FormsModule } from '@angular/forms';

import { ButtonComponent } from './button.component';
import { ButtonType } from './button-type';

describe('Button component - ', () => {
  let comp: ButtonComponent;
  let fixture: ComponentFixture<ButtonComponent>;

  beforeEach(() => {
  });

  beforeEach(async(() => {
    TestBed.configureTestingModule({
      imports: [
        BrowserAnimationsModule,
        FormsModule
      ],
      declarations: [
        ButtonComponent
      ]
    })
      .compileComponents()
      .then(() => {
        fixture = TestBed.createComponent(ButtonComponent);
        comp = fixture.componentInstance;
        fixture.detectChanges();
      });
  }));

  it('should have a default type', () => {
    let button = fixture.debugElement.query(By.css('.pf-c-button'));
    expect(button).not.toBeNull();
  });

  it('should have type danger', () => {
    comp.type = ButtonType.DANGER;
    fixture.detectChanges();

    let button = fixture.debugElement.query(By.css('.pf-c-button.pf-is-danger'));
    expect(button).not.toBeNull();
  });

  it('should have type primary', () => {
    comp.type = ButtonType.PRIMARY;
    fixture.detectChanges();

    let button = fixture.debugElement.query(By.css('.pf-c-button.pf-is-primary'));
    expect(button).not.toBeNull();
  });

  it('should have type secondary', () => {
    comp.type = ButtonType.SECONDARY;
    fixture.detectChanges();

    let button = fixture.debugElement.query(By.css('.pf-c-button.pf-is-secondary'));
    expect(button).not.toBeNull();
  });

  it('should have type tertiary', () => {
    comp.type = ButtonType.TERTIARY;
    fixture.detectChanges();

    let button = fixture.debugElement.query(By.css('.pf-c-button.pf-is-tertiary'));
    expect(button).not.toBeNull();
  });

  it('should call a function when button is clicked', () => {
    let event: any = null;
    comp.onSelect.subscribe(($event: ButtonComponent) => {
      event = $event;
    });
    let button = fixture.debugElement.query(By.css('.pf-c-button'));
    expect(button).not.toBeNull();

    button.nativeElement.click();
    fixture.detectChanges();
    expect(event).not.toBeNull();
  });
});
